requirements.txt
requests==2.32.5
Flask==3.1.2
Flask-Cors==6.0.1
Flask-JWT-Extended==4.7.1
pandas==2.3.3
python-dotenv==1.2.1
psycopg[binary,pool]==3.2.12
gunicorn==23.0.0